Changed open_rate from Int to Double

+++ src/main/java/com/ecwid/maleorang/method/v3_0/lists/StatsInfo.kt
@@ -47,23 +47,23 @@ class StatsInfo : MailchimpObject() {
 
     @JvmField
     @Field
-    var avg_sub_rate: Int? = null
+    var avg_sub_rate: Double? = null
 
     @JvmField
     @Field
-    var avg_unsub_rate: Int? = null
+    var avg_unsub_rate: Double? = null
 
     @JvmField
     @Field
-    var target_sub_rate: Int? = null
+    var target_sub_rate: Double? = null
 
     @JvmField
     @Field
-    var open_rate: Int? = null
+    var open_rate: Double? = null
 
     @JvmField
     @Field
-    var click_rate: Int? = null
+    var click_rate: Double? = null
 
     @JvmField
     @Field
